Bump zcash_primitives to 0.8
