+++ ledger-zcash-builder/Cargo.toml
@@ -47,7 +47,7 @@ secp256k1 = { version = "0.29" }
 
 #zcash
 wagyu-zcash-parameters = { version = "0.2", optional = true }
-
+ledger-zcash-legacy-helpers = { path = "../ledger-zcash-legacy-helpers" }
 zcash_primitives = { version = "0.8", features = ["transparent-inputs"] }
 zcash_proofs = { version = "0.8", features = ["multicore"], optional = true }
 zcash_note_encryption = { version = "0.1", features = ["pre-zip-212"] }
