+++ ledger-zcash/src/txbuilder.rs
@@ -35,7 +35,7 @@ use zcash_primitives::{
 use zx_bip44::BIP44Path;
 
 use crate::{DataInput, DataShieldedOutput, DataShieldedSpend, DataTransparentInput, DataTransparentOutput, ZcashApp};
-
+use ledger_zcash_legacy_helpers::script_to_address;
 /// Ergonomic ZCash transaction builder for HSM
 #[derive(Default)]
 pub struct Builder {
@@ -61,8 +61,8 @@ impl TryFrom<DataInput> for Builder {
 
         for to in input.vec_tout.into_iter() {
             builder.add_transparent_output(
-                &to.script_pubkey
-                    .address()
+                
+                    &script_to_address(&to.script_pubkey)
                     .ok_or(BuilderError::InvalidAddress)?,
                 to.value,
             )?;
@@ -187,7 +187,7 @@ impl Builder {
             ripemd.finalize()
         };
 
-        match coin.script_pubkey.address() {
+        match script_to_address(&coin.script_pubkey) {
             Some(TransparentAddress::PublicKey(hash)) if hash == pkh[..] => {},
             _ => return Err(BuilderError::InvalidUTXOAddress),
         }
@@ -520,7 +520,7 @@ impl Builder {
     where
         R: RngCore + CryptoRng,
         TX: HsmTxProver + Send + Sync,
-        P: Parameters + Send + Sync,
+        P: Parameters,
         E: ledger_transport::Exchange + Send + Sync,
         E::Error: std::error::Error,
     {
